Update README.md
An integrated website for selling fashion. There is protection for logging in and the password. Once entered, there is protection to encrypt it to protect customer data. There is an admin with a username and password to add products and their prices.

admin@email.com / username
admin123123 / password


There is no login or password for the database, but there is an admin who can access the database. Above it is written the password, username, and email.